fix: align type names with JSON Structure Core spec
Remove non-spec type names (float16, float32, float64, float128,
decimal64, decimal128) from all SDKs. The spec defines:
- float8: 8-bit floating point
- float: 32-bit single precision (IEEE 754 binary32)
- double: 64-bit double precision (IEEE 754 binary64)
- decimal: arbitrary precision decimal

.NET SDK:
- SchemaValidator.IsNumericType: use spec types
- InstanceValidator: remove non-spec case labels
- JsonStructureSchemaExporter: export float/double instead of float32/float64
- Update test expectations
- Add demo app for error message visualization

Java SDK:
- SchemaValidator: fix PRIMITIVE_TYPES and NUMERIC_TYPES
- InstanceValidator: fix switch statement
- JsonStructureSchemaExporter: emit float/double
- Update test schemas and assertions
- Update README documentation

+++ dotnet/src/JsonStructure/Schema/JsonStructureSchemaExporter.cs
@@ -357,10 +357,11 @@ public static class JsonStructureSchemaExporter
         if (type == typeof(ulong)) return "uint64";
         if (type == typeof(UInt128)) return "uint128";
 
-        // Floating point
-        if (type == typeof(Half)) return "float16";
-        if (type == typeof(float)) return "float32";
-        if (type == typeof(double)) return "float64";
+        // Floating point - use spec names: float (32-bit), double (64-bit)
+        // Note: Half (16-bit) maps to float as the closest available spec type
+        if (type == typeof(Half)) return "float";
+        if (type == typeof(float)) return "float";
+        if (type == typeof(double)) return "double";
         if (type == typeof(decimal)) return "decimal";
 
         // Date/time types
